Перенос вспомогательные функции в util
