+++ js/data.js
@@ -1,6 +1,12 @@
 import { getRandomNumb } from './util.js';
 import { getMockData } from './mock.js';
 
+const MAX_PRICE = 1000000;
+const MAX_ROOMS = 30;
+const MAX_GUESTS = 100;
+const MAX_PHOTOS = 10;
+const COORD_DECIMALS_SIGNS = 5;
+
 const tempData = getMockData();
 
 const getAuthor = () =>  ({ avatar: 'img/avatars/user0' + getRandomNumb(1, 8) + '.png' });
@@ -24,8 +30,8 @@ const getPhotos = (photosArr, maxPhotos) => {
 
 const getLocation = () => {
   return {
-    x: getRandomNumb(tempData.latitudeMin, tempData.latitudeMax, 5),
-    y: getRandomNumb(tempData.longitudeMin, tempData.longitudeMax, 5),
+    x: getRandomNumb(tempData.latitudeMin, tempData.latitudeMax, COORD_DECIMALS_SIGNS),
+    y: getRandomNumb(tempData.longitudeMin, tempData.longitudeMax, COORD_DECIMALS_SIGNS),
   }
 }
 
@@ -33,15 +39,15 @@ const getOffer = (AddressCoord) => {
   return {
     title: tempData.tempTitle,
     address: Object.values(AddressCoord).join(', '),
-    price: getRandomNumb(0, 100000),
+    price: getRandomNumb(0, MAX_PRICE),
     type: tempData.typesOfDwelling[getRandomNumb(0, tempData.typesOfDwelling.length-1)],
-    rooms: getRandomNumb(0, 10),
-    guests: getRandomNumb(0, 30),
+    rooms: getRandomNumb(0, MAX_ROOMS),
+    guests: getRandomNumb(0, MAX_GUESTS),
     checkin: tempData.checkTimes[getRandomNumb(0, tempData.checkTimes.length-1)],
     checkout: tempData.checkTimes[getRandomNumb(0, tempData.checkTimes.length-1)],
     features: getFeatures(tempData.featuresList),
     description: tempData.tempDescription,
-    photos: getPhotos(tempData.photosList, 10),
+    photos: getPhotos(tempData.photosList, MAX_PHOTOS),
   }
 }
